Fix alt page links and export helper
