+++ test/cloud-functions/buildAltsHtml.test.js
@@ -1,0 +1,9 @@
+import { describe, test, expect } from '@jest/globals';
+import { buildAltsHtml } from '../../infra/cloud-functions/render-variant/buildAltsHtml.js';
+
+describe('buildAltsHtml', () => {
+  test('links to variant pages in /p path', () => {
+    const html = buildAltsHtml(1, [{ name: 'a', content: 'hello world' }]);
+    expect(html).toContain('<a href="/p/1a.html">');
+  });
+});
